ZUGFeRD — split ZUGFeRD invoice-PDF E2E into own spec; run en_US + de_DE
Per the 'one createMasterdata per spec file' rule: move the ZUGFeRD case out of
shipment.spec.js (back to pure O2C) into zugferd-invoice.spec.js, reusing the same
SalesOrderPage/ShipmentSchedulePage/InvoiceCandidatePage/InvoicePage flow. Run it in
both en_US and de_DE (language-invariant selectors). No IsMockReportService — the real
report service renders the PDF/A-3 that the archive seam embeds the CII into.
